# src/main/resources/openapi/v1/order-openapi.yaml
openapi: 3.1.0
info:
  title: Order Service API
  description: |
    REST API for managing customer orders in an e-commerce platform.
    
    Key features:
    - Place new orders (from authenticated user's cart)
    - Retrieve order lists (all / per user)
    - Get detailed order information
    - Mark order as paid (internal/admin)
    - Cancel orders
    - Basic health check
    
    Authentication: JWT Bearer token required for most endpoints.
    User ID is typically extracted from the JWT subject (`sub` claim).
  version: 1.0.0
  contact:
    name: Infinity Tech Support
    email: infinitytech@support.co.za
  license:
    name: Proprietary License - All rights reserved.
    url: https://infinitytech.co.za/terms

servers:
  - url: https://api.infinitytech.com/v1
    description: Production server
  - url: https://staging-api.infinitytech.com/v1
    description: Staging server

tags:
  - name: Orders
    description: Operations related to creating, viewing, updating status and cancelling orders
  - name: Health
    description: Service health and monitoring endpoints

paths:
  /api/v1/order/health:
    get:
      tags:
        - Health
      summary: Check if the order service is healthy
      operationId: healthCheck
      responses:
        '200':
          description: Service is operational
          content:
            text/plain:
              schema:
                type: string
              example: "Order service is up and running!"
  
  /api/v1/order/all:
    get:
      tags:
        - Orders
      summary: Get all orders (admin / support view)
      description: Returns a list of all orders in the system. Usually restricted to admin roles.
      operationId: getAllOrders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  
  /api/v1/order/place-order:
    post:
      tags:
        - Orders
      summary: Place a new order for the authenticated user
      description: |
        Creates a new order based on the authenticated user's current cart contents.
        User ID is extracted from JWT subject claim.
        Typically called after cart finalization and payment initiation.
      operationId: placeOrder
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order successfully placed and queued for processing
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '400':
          description: Invalid cart state / empty cart / address missing
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Conflict (e.g. stock changed since reservation)
  
  /api/v1/order/mark-paid/{orderId}:
    put:
      tags:
        - Orders
      summary: Mark an order as paid
      description: |
        Updates order status to PAID.
        Usually called by payment service / webhook after successful payment.
        Restricted to internal services or admin roles.
      operationId: markOrderAsPaid
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order marked as paid
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Order not found
        '409':
          description: Order already paid / cancelled / in invalid state
  
  /api/v1/order/{orderId}:
    get:
      tags:
        - Orders
      summary: Get detailed information about a specific order
      operationId: getOrderDetails
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User not authorized to view this order
        '404':
          description: Order not found
  
  /api/v1/order/user/{userId}:
    get:
      tags:
        - Orders
      summary: Get all orders for a specific user
      description: Returns order history for the given user ID
      operationId: getUserOrders
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: User identifier (usually UUID from JWT sub)
      responses:
        '200':
          description: List of user's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not authorized to view this user's orders
        '404':
          description: User not found (optional)
  
  /api/v1/order/{orderId}/cancel:
    delete:
      tags:
        - Orders
      summary: Cancel an order
      description: |
        Cancels an order if it is in a cancellable state (e.g. not yet shipped).
        Triggers stock release and notification.
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Order successfully cancelled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Order cannot be cancelled (already shipped, etc.)
        '404':
          description: Order not found

components:
  schemas:
    OrderDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10045
        userId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        orderNumber:
          type: string
          example: ORD-20260304-10045
        status:
          type: string
          enum: [PENDING, PAID, PROCESSING, SHIPPED, DELIVERED, CANCELLED, REFUNDED]
          example: PAID
        totalAmount:
          type: number
          format: decimal
          example: 2499.95
        currency:
          type: string
          example: ZAR
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            type: object
            properties:
              productId: { type: integer }
              name: { type: string }
              quantity: { type: integer }
              unitPrice: { type: number }
              subtotal: { type: number }
        shippingAddress:
          type: object
          properties:
            street: { type: string }
            city: { type: string }
            postalCode: { type: string }
            country: { type: string }
        paymentMethod:
          type: string
          example: CARD
      required:
        - id
        - userId
        - status
        - totalAmount
    
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        detail:
          type: string
          nullable: true
  
  responses:
    BadRequest:
      description: Validation error or invalid request state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT